requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pandas>=2.1.4
nba_api>=1.4.1
python-dotenv>=1.0.0
xgboost>=2.0.0
scikit-learn>=1.3.1
joblib>=1.3.0
toml>=0.10.2
